alsa-state: Rename init script
Having a SRC_URI called ${PN} is asking for trouble. When extending FILESPATH,
alsa-state can be treated as a directory and copied over the contents of ${WORKDIR}
which is invariably not what the user wants.

Avoid this by renaming the SRC_URI to something else and only call
it alsa-state at install time.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>