oe-core/kernel: Deploy KERNEL_ALT_IMAGETYPE and check it for uImage
Install, deploy and package KERNEL_ALT_IMAGETYPE and create kernel-alt-image
package.

Check if KERNEL_ALT_IMAGETYPE is empty or not. If KERNEL_ALT_IMAGETYPE is empty,
just skip alt image install and deploy, otherwise it is failed to install/deploy
because KERNEL_ALT_OUTPUT doesn't point to a real kernel image file.

Check if KERNEL_ALT_IMAGETYPE is uImage in do_uboot_mkimage and adding
u-boot-mkimage-native depend if it is uImage.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20140716

Signed-off-by: Yang Shi <yang.shi@windriver.com>