bluez5: fix QA error from configure option
Issue: LINNT7-62

(master rev: 8c9c08112d7b3bde48fb4dd5e15429948aca0dd6)

The systemdunitdir option was split into systemdsystemunitdir and
systemduserunitdir before bluez5 was ever released, so this produced a
QA error and was ignored.  There appears to be no reason to override the
inferred default, so replace it with an explicit --enable-systemd.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Jackie Huang <jackie.huang@windriver.com>