eglibc-use-option-groups.patch: Minor fixups.
This incorporates suggested changes for issues 6797 and 6809 from
bugzilla, and a revised change for 6796. For 6796, the issue is
that you can't just include <gnu/option-groups.h> in libc-symbols,
so we may not always have the option group defines available.
If it's always visible, glibc won't build for poky-tiny, because
some code assumes that it can link against locale code if it
finds HAVE_MBSTATE_T. So, for now, test defined() before testing
the value of the option group feature test macro. It's possible
that this will result in some code being compiled without
HAVE_MBSTATE_T when it should in cases where option-groups.h wasn't
included before libc-symbols.h, but I haven't found an actual
example of a failure.

The reason HAVE_MBSTATE_T wasn't conditionalized in 2.19 is that
it didn't exist in 2.19.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20141111

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>