image_types_uboot: add the third parameter for function oe_mkimage
Issue: LIN7-3179

The previous fix for LIN7-3179 is removing the file passed into
oe_mkimage directily, that causes do_rootfs failed in some cases,
because sometimes, the file passed into oe-mkimage must be kept.

So the new fix is adding the third parameter for oe_mkimage to control
removing or not removing.

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>