Improve the regex of sed command of SVN_SERF_INCLUDES macro
Issue: LIN7-3989

The regex of sed command is too simple, which mangles any "-D"
in strings, for example, a project name is "My-Demo", it will
be removed from the path by the sed regex, which causes a build
failure. Improving it to only delete the -D behind of a space or
the -D at start of a string.

Reference

ttps://bugzilla.yoctoproject.org/show_bug.cgi?id=7874

Upstream-Status: Pending

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>