rpm: check if the argument(rootpath) exists or be writable
Issue: LIN7-3977, LIN7-3997

When user execute the command "rpm -qai --root=$dir",if $dir doesn't
exist or is unwritable as result of making a typo in rootpath,then
it will create dirent $dir and subdirectory.
So we should add the check function to fix it before creating relational
subdirectory,and warn the incorrect rootpath to user. It just checks the
rootpath reasonableness when the user input the argument(--root=/-r=).

(LOCAL REV: NOT UPSTREAM) -- pending

Signed-off-by: Zhixiong Chi <zchi@windriver.com>