binutils: fix regression in CVE-2014-8737 fix [ LIN7-4690 ]
Issue: LIN7-4690

TCWRLGCC-15/LIN7-2439 made a mistake in backporting upstream patch and left
what should be deleted in source code. It results in a regression on objcopy.
Error message is shown when copying archive files regardless of the validness
of path. This fix addresses this regression.

Upstream-status: Inappropriate [other]
The patch being dependent on has not sent to OE yet.

Signed-off-by: Yuanjie Huang <yuanjie.huang@windriver.com>