openssl: don't move libcrypto to base_libdir
Issue: LIN8-961

For now, if 'openssl' is enabled for ntp, ntp would still be built
without openssl & libcrypto. This is because that ntp thinks openssl
and libcrypto locates under the same directory.

A patch more than two years ago moves libcrypto from libdir to base_libdir
for the reason of /usr split. And I don't think this reason still stands.

This patch removes the codes of moving libcrypto to base_libdir.

(LOCAL REV: NOT UPSTREAM) -- sent to oecore on 20151016

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>