gnupg: set development_version as no
Issue: LIN8-1149

The development_version used to mark if gnupg is a beta version. The
checking logic in autogen.sh always assume that gnupg is beta if below
command is not run in gnupg's git-repo:

  './autogen.sh --find-version 2 1 7'

This does not work for integration. Set development_version as no to
fix runtime issue:

  $gpg --list-sigs
  gpg: NOTE: THIS IS A DEVELOPMENT VERSION!
  gpg: It is only intended for test purposes and should NOT be
  gpg: used in a production environment or with production keys!

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151022

Signed-off-by: Wenzong Fan <wenzong.fan@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>