coreutils: fix reporting 'unknown' by `uname -p' and `uname -i'
Issue: LIN8-1308

This patch make `uname -p' and `uname -i' not reporting 'unknown'.
It refers a to Fedora's way to do this.

The coreutils upstream rejects to accept this patch, blaming the
'unknown' result to the kernel not providing enough information.
As on normal distros, `uname -p' and `uname -i' do not report 'unknown',
we refer to Fedora's patch to fix this problem.

(LOCAL REV: NOT UPSTREAM) -- sent to oecore on 20151029

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>