Use TUNE_CCARGS for target compilation
ISSUE: LIN8-1048

The initial gcc configuration is using TARGET_CFLAGS, but this is
probably not what is intended, because TARGET_CFLAGS doesn't include
any architecture-specific flags.

For PPC64, it's possible to end up with gcc configured to pass -a64 to
the assembler, but be itself getting invoked without -m64, which can result
in alignment problems in generated code. Solution: Instead of setting
CFLAGS_FOR_TARGET to TARGET_CFLAGS, set it to TARGET_CFLAGS + TUNE_CCARGS.

This is not fully validated, but resolves the specific failure we've
been seeing, and sort of makes sense.

Upstream-status: Not submitted yet, but probably should be.

(LOCAL REV: NOT UPSTREAM) -- will send to oe-core

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>