package_manager.py: Avoid installing an empty package list
It is possible in an attempt only install, that everything listed is not
available to be installed.  This will have the effect of clearing the
package list.  However, we only check for an empty package list at
the beginning of the function.  We need to also check before running the
install, otherwise we can fail due to 'error: no package(s) given".

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151125

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>

Issue: LIN7-3328

Signed-off-by: Yuanjie Huang <yuanjie.huang@windriver.com>