allarch.bbclass: allarch packages are having their hashes altered by baselib
In a multilib configuration baselib can change, but allarch packages should not
changed from multilib to multilib.  So we set the value of baselib to none
which should act as a market for any errant users of libdir.
nonarch_base_libdir / nonarch_libdir should be used instead.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151125

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>