kernel: Prevent hash from depending on KERNEL_VERSION
Issue: LIN8-1692

70e5926828da25d625282c0426d98409364be1b9 "kernel: Add KERNEL_IMAGETYPES
to build multi types kernel at one time" reads the value of
'KERNEL_VERSION', immediately evaluates it and uses it to set other
variables. The immediate evaluation of KERNEL_VERSION is based on a
dynamic value. Until the kernel is extracted and built it returns 'None'.

So the result is the first time the kernel is built the hash gets 'None'
embedded in it. The second time it's build it gets the actual kernel
version. This triggers a rebuild condition.

This patch:
 - Make the hash depend on KERNEL_VERSION_PKG_NAME which depends on
   LINUX_KERNEL's value which is constant
 - Removes the unnecessary replacements of KERNEL_VERSION and PV in
   kernel-grub.bbclass

(LOCAL REV: NOT UPSTREAM) -- will send to oe-core on 20151130

Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>