rpm_5.4.14: Enable MIPS64 N32 transactions
Issue: LIN8-1977

In a multilib install RPM will break a conflict using the transaction color.
However, by default MIPS64 N32 transactions are not enabled.

This change always enables them.  This is effectively a no-op on non-MIPS
systems, and required if we need to support N32 on MIPS.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151202

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>