cross-canadian.bbclass: Add BASECANADIANEXTRAOS to specify main extraos
Issue: LIN8-2098

By default the system will expand the extra os entries for uclibc and musl
even if they are not enabled in the build.  There was no way to prevent this
behavior while still getting the expansion for things like x32 or spe.

The change adds a new setting which a distribution creator can override
easily, setting the base set of canadianextraos components.  The other
expansions are then based on this setting.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151207

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>